Allows simple "import { Vec3 } from 'vec3';" in ES modules.
Adds an ESM wrapper, the generator for the wrapper as a dev dependency, and adds a prepack trigger to ensure the wrapper is kept up to date with the actual exports of the CJS version of the code. Using a wrapper helps prevent the dual package hazard, as would otherwise occur if the code was written as ESM and completely transpiled to CJS (or vice versa) and given a second entry point (instanceofs between CJS and ESM imports of the code would fail).